refactor: the homepage layout
