+++ src/views/Faq/index.tsx
@@ -21,7 +21,7 @@ function Faq() {
           >
             {t('home.openEditor')}
             {' '}
-            ➡️
+            ➜
           </Link>
         </div>
