+++ src/views/Home/components/WallOfLove.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable react-dom/no-missing-iframe-sandbox */
 import { useEffect } from 'react'
+import { useTranslation } from 'react-i18next'
 
 function WallOfLove() {
   useEffect(() => {
@@ -8,11 +9,20 @@ function WallOfLove() {
     script.setAttribute('src', 'https://widget.senja.io/js/iframeResizer.min.js')
     document.body.appendChild(script)
   })
+  const { t } = useTranslation()
+  const tweetText = encodeURIComponent(t('home.tweetText'))
 
   return (
-    <div>
+    <div className="md:h-[800px] h-screen w-full">
+      <button type="button" className="flex mx-auto mb-4 hover:translate-x-2 duration-300 bg-blue-400 rounded-full px-6 text-white text-xl font-Inter font-semibold p-2">
+        <a href={`https://twitter.com/intent/tweet?text=${tweetText}`} rel="noreferrer" target="_blank">
+          {t('home.shareOnTwitter')}
+          {' '}
+          👏
+        </a>
+      </button>
       <iframe
-        className="w-9/12 h-screen  mx-auto border-none overflow-hidden"
+        className="md:w-9/12 h-full py-4 px-6 mx-auto border-none"
         width="100%"
         height="100%"
         id="senja-frame-902012ea"
